# back-end/local-app/src/main/java/br/com/novaalianca/mnss/localapp/domain/catalog/ProductController.java
package br.com.novaalianca.mnss.localapp.domain.catalog;

import br.com.novaalianca.mnss.localapp.security.auth.AuthenticatedUser;
import org.springframework.security.access.prepost.PreAuthorize;
import br.com.novaalianca.mnss.localapp.security.user.RoleName;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.validation.Valid;
import java.util.List;
import java.util.UUID;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/products")
class ProductController {
    private final CatalogService catalogService;

    ProductController(CatalogService catalogService) {
        this.catalogService = catalogService;
    }

    @GetMapping
    List<ProductResponse> listProducts(
            @RequestParam(required = false) String name,
            @RequestParam(required = false) UUID categoryId) {
        return catalogService.listProducts(name, categoryId);
    }

    @GetMapping("/barcode/{barcode}")
    ProductResponse findProductByBarcode(@PathVariable String barcode) {
        return catalogService.findSellableProductByBarcode(barcode);
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    @PreAuthorize("hasRole('GERENTE')")
    ProductResponse createProduct(
            @Valid @RequestBody CreateProductRequest request,
            HttpServletRequest servletRequest) {
        return catalogService.createProduct(request, authenticatedUserId(servletRequest));
    }

    @PatchMapping("/{id}")
    @PreAuthorize("hasRole('GERENTE')")
    ProductResponse updateProduct(
            @PathVariable UUID id,
            @Valid @RequestBody PatchProductRequest request,
            HttpServletRequest servletRequest) {
        return catalogService.updateProduct(id, request, authenticatedUserId(servletRequest));
    }

    @PatchMapping("/{id}/availability")
    @PreAuthorize("hasAnyRole('GERENTE', 'ATENDENTE')")
    ProductAvailabilityResponse updateProductAvailability(
            @PathVariable UUID id,
            @Valid @RequestBody PatchProductAvailabilityRequest request,
            HttpServletRequest servletRequest) {
        return catalogService.updateProductAvailability(id, request, authenticatedUserId(servletRequest));
    }

    private UUID authenticatedUserId(HttpServletRequest request) {
        Object attribute = request.getAttribute(AuthenticatedUser.AUTHENTICATED_USER_ATTRIBUTE);
        return attribute instanceof AuthenticatedUser user ? user.id() : null;
    }
}